requirements.txt
aiogram==3.4.1
asyncpg==0.29.0
python-dotenv==1.0.0
alembic==1.13.1
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
requests==2.31.0
psutil==5.9.6 